First Change
This is the first big change in the project from the base. The problem is whether I will keep this or not and maybe just change some things up back on the Mac rather than import it from github. Anyways the problem is the fact that I moved it from Mac to Windows so some line endings have changed and stuff but not just that, I also used it in a new version of Unity and on monodevelop rather than Visual Studio since I didn't  have time to download the new .NET framework but we'll see. Idk if I succesfully made a base and uploaded the proper changes, if not that it is fine because I have both project files and can just restart this repository from scratch.