2nd Update
This one is included with a more refined script, control, and system of ground detection. I just improved and changed some things in the script as shown. I also want to create an animator and start with level mechanics. As much as I want to make more gmaeplay elements and mechanics I have to start with at least one REAL level. Greg is almost done with the enemy sprite then he's moving on to other sprites. With the enemy sprite I will be able to make enemies and have one basic level.