3rd Update
Today I finally worked on adding animations and fitting the sprites into new dimensions. I just added an animation controller to the character game object and added two animations to the rpoject with just the jumping and falling left(which are only one frame each). Next I hope to finish those and get them synced up to the script and making some of the levels. After that I may finalize some game mechanics and make enemies. Just some animations and fixes to the dimensions of each sprite (20x19 now (or 19x20 idk)).