4th Update
This time I finished the animation imports and hooked them up to code. So faralmost everything is working except for the falling animation which I may have to tweek or just redo completely so it works right. I shifted the groundCheck a bit to get it lined up how I want it to. I messed with some animation to make them faster such as with falling and jumping and just hooked up the animator to the script so everything was fine. That's it, once I finalize the falling animation then I will begin on camera movement as well as a first level. Meanwhile Greg has been creating the enemy sprite which I will have to code sometime(aw man) and I will also have to code health and such. I also need a display for the science fair which is gonna take more time out of this project. This is gonna be a commit that I will push tonight and I might push another smaller one later tonight if I have the time. If it's smaller it will just be an addon to this update but if I devide to go all out and start the camera stuff and its big, then it will be update 5.