Update 5.0
Finished fixing the falling animation and everything looks nice. I added a suitable camera and changed the speed to make it faster and made the camera closer. Not much other than a new camera script and a new bool added in the animator and the script to help control everything. I changed the level a bit so I could make a longer runinning scape and some more freedom. Next I hope to work on a level and intigrate an enemy to it to get at least something.