Update 6.0
Huge update and I'm hoping to work more tonight. So I just imported some hearts and made a lives counter that works and everything and can be alligned with any amount of lives. Took some time to make it though with some errors but everything is fine. I changed the ground a bit to accomodate more for testing different things. I started on an enemy and hoping to finish it by tonight so then I can work on the menu tomorrow then finally just make 5 or so levels. I added a background that is always sationary unfortunately. I added the game manager for future use. I just moved some files around and made a prefab for the heart which is used for the GUI to showcase how many lives a player has. There's probably more but you(future me) can see them more specifically in each one.