Update 6.5
This isn't fully a big enough update to be a full numerical value higher than the last one but it is pretty significant with a lot of files and a lot of time put into it that it deserves a higher spot than 6.1 so I settled with 6.5. I just worked a lot on the enmy and creating some behaviour from it. Mainly I made the collisions work mostly properly so when he gets jumped on he dies and he deals damage and he attacks when in the player gets into a trigger zone. It took some time to get the rotation working but I just settled on something that mostly worked. Sometimes he would spazz out and I think I have a way of fixing it but not right now, but I know exactly why it happens I'm just too exhausted and tired. I created two animations and some control over them so it would look at least like something. I messed around with the Physics2D stuff but I have a good enough template for an enemy which I just need to make a prefab of when i can. I also need to add invincibility frames when he gets hurt and fix something where his hearts don't seem to dissappear when he gets hit by the enemy. I also need to adjust some things when I get Greg's new spritesheet working which I will need to change the animations for the enemies and if I have to reimport everything fresh then that means a lot of time for recreating animations and taking out each frame. Yipee.