Update 6.7
New day but not much done. All I did was fix some code in the enemy character but not enough. I changed the radius when player is detected and I also fixed the heart placement and if lives deplete when getting hit. I still have to fix some inconsistencies to fix but for the most part it works. He goes invincible when he gets hit and I'm working on making him flash a bit and things like that. More to do later today.