Update 6.8
I should have labeled the lst update Update 6.5.1 or something like that. But anyways update 6.8 is rolling out and some new stuff got added, and enough to make a difference; also I just found out Mathf.PingPong was a thing. First of all, I shifted the colliders(specfically the capsule collider) to be smaller and fit right inside of the box collider for the palyer. This also was accompanied by me changing the size of the box for the Physics.2D thing. I adjusted it so that hopefully it would get rid of the walljumping problem without removing the area to jump if a person got stuck on a corner. Another change was the layout of the map. I just shifted everything over and moved the positions of the enemy and player. Also I fixed the rotation of the enemy when it is facing a player. On top of that I fixed collisions so some colliders would be ignored to make the damaging of the player more consistent and jumping on top of the enemy(so now you don't get damaged accidentally when jumping on the enemy). Also on the damage, I made it so now the player blinks when invincible(therefore deleting the old animation fail which I think is the cause and still is the cause for a weird error I get idk) and made sure that collisions were still active when the scene was reloaded. The scene now reloads when the plaayer dies(loses all lives). I made it so now the player starts next to the wall and has to press jump before he starts moving. This can shocase the idle animation and help control the transitioning of levels. I changed the cameraso that it showed more room towards the way the player was facing. So whenever the player switched directions, the camera would readjust to show more of what is ahead of the player(im going to play around with the value before it is finalized). I think that is all from what I see has changed(there may have been some things I missed) but soon I'm hoping to just crunch out 5 levels, and when Greg is finished with the sprites I can add the nimations to the enemies and stuff. There is still a lot of readjusting to be made later on. That's it for now.