Update 7.0
I'd say this is worthy of a higher number for the update. There was enough stuff that was changed and added to grant this title. I added the heart tag to the prefab so that I could more easily recognize it when cycling through children. Also I changed a bit of code in that method to make it look for the tags of the children rather than of the player. I added the final stuff to the death of the player. Now instead of restarting a level, a text and button appear to indicate that you have to restart the level(those seperate game objects were created). I made it so you get shot backwards by the enemy when hit and that you spin out when dead to make the death a little better looking.  I also added a text box that tells you to press the "" key to start. I made it so the camera stops following the player after they die. In the end I added a main menu with the title and our names as well as a play button to start the game. All commands done by th buttons can also be done with space bar. And it moves to the next scene according to the build index which I set up. That is most of everything very simplified but I think that the details of this update really make it deserve something higher than an increase in the tenths place and rather in the whole numbers.