Update 8.0
Wow, this is literally the last haul. So much in one night that it is hard to break down everything. So look at the files for specific changes since I am not going to list them  all. Although here are a few honorable mentions: First off, I tried to implement Greg's enemy death animation but it wasn't possible, and not just because it was a drastically different size than the formal sprites, but also the fact that Unity didn't even play it anyway for some reason. So I just made the enemy bounce up and fall out of frame. I changed the lives counter to work better and gave the character working code for an extra life when I do implement it. I added kill planes for bottomless pits and a next level trigger for advancing to the next level. I changed the idle to include the normal sprite for the enemy. I created an enemy prefab and player prefab for reuse when I need it. Alright, last and certainly not least, the levels. I finished the first level with it being short and everything and only one enemy. But then I started and almost completed the second level. I just have to test a few jumps and add the kill planes and I'm all set. The second level is very tough and has a good amount of enemies. Tomorroow is a day to work on the data notebook and create and finish the third level. It shouldn't be too hard to create if I keep it to a minimum. This update was huge(I was almost contemplating making it 9.0 because of the size) which means there are a lot more details in the changelog and I just simplified a few things and didn't mention some stuff yet. I was hoping to finish today but I should certainly be finished by tomorrow hopefully.