Update 9.0
Huge update, but probably last update for a long time. This has been really stressful so I think it will be nice to take a break. Ok so this is the build that will be used for the Engineering project which means that it's not the most polished but it gets the job done. Ok so to summarize another big update: I finished the second level and also made and finished a third level. For the second level i just tweeked some of the platforms and enemy placements. The third level is just a path leading out of the cave. There is a bit of platforming in the beginning then there's some travelling updwards and then the last bit of platforming with some enemy placements and finally like 10 enemies outside and the finish of the game(a trigger). I added tags for kill boxes, end level triggers, and the ending for the game. All of which the game objects were created except for the end game one. I created that along with a little ui to pop up that will send you back to the menu. I messed around with the extra lives, I added the game object and added some code for the player whenever he picks one up. I changed the sorting order of some things, but specifically so I could have the line renderer show up on top of one thing but still in the back. Iadded two line renderers in two levels, one in each. These are just blue line renders that act as background just to cut corners and are used for the underground portion. I made it so the player gets pushed back less when hit but pushed back a lot when dead for just looks. Also I made it so his gameobject gets distroyed at a certain point but idk if it will work. And mainly overall, I just twweked some gameobjects in the editor and levels so they would fill up their spots in references and such. And lastly i just fixed player seetings so the Game would be ready for building and be finished(which was done).